requirements.txt
click>=8.1.0
python-dateutil>=2.8.2
tabulate>=0.9.0
rich>=13.3.0
google-generativeai>=0.3.0
